requirements.txt
Flask>=0.12.1
Jinja2==2.9.6
cf-deployment-tracker==1.0.4
numpy==1.14.0
cloudant==2.4.0
requests==2.12.4
watson-developer-cloud>=1.2.1
python-twitter
